Refactor duplicated code for Game page.
