+++ _templates/global/Game.html
@@ -2,8 +2,9 @@
     <link href="{% static 'sliders.css' %}" rel="stylesheet"/>
 {% endblock %}
 
+
 {% block title %}
-Stage 2: Round {% player.round_number %} of 5
+Round {% player.round_number %} of 5
 {% endblock %}
 
 {% block content %}
