+++ sliders/image_utils.py
